Add gui modules
The two new files pyvsm-gui.py and vsmutils.py contain classes to construct
a GUI frontent for the analysis routines in pyvsm.py. I started writing
this about two weeks ago, and since it's going well so far, I decided
it's time to add it to the pyvsm repo.

In addition to facilitating the use of pyvsm for other Newman Group
members, this is a chance for me totry building a fairly comprehensive
graphical interface using Tk. I have been learning as I go, and will
continue to do so.