Extend Sample Manager Buttons
In addition to the 'add sample' button below the sample list box,
there are now 'rename sample' and 'remove sample', which allow the user
further control over the sample list.

Another feature to consider would be buttons to move the current working
sample up or down in the list. This would have no functionality at the
moment, but might later if the order of samples takes on some significance.

Adding the 'remove sample' button has uncovered the fact that most
functions that use the current working sample do not handle the possibility
of there not being a current working sample, which would happen if the user
removed all samples. Handling this case is next on my to-do list.