Handle case of no samples
Each method that requires use of the session's current working sample
has been modified to handle the possibility of there not being a current
working sample. This case should only arise when the user has deleted all
samples. Generally, the case is handled by checking if
session.currentWorkingSample evaluates to True. This assumes that
session.currentWorkingSample being unusable also implies that it will
evaluate to False, which I can't say for sure is true. It may be more
robust to catch exceptions as they come up.