Add results text
There is now a text box which displays the contents of the results
dict created by VsmAnalyzer.analyzeData. I wanted the user to be able
to highlight and copy the text, but to otherwise be unable to edit it,
similar to the 'readonly' mode of the tkinter.Entry widget. the
tkinter.Text widget does not support this behavior by default, so a
child class, ReadOnlyText, has been created (in readonlytext.py), which
does.

Currently the resutls are printed as name: value. This is all right,
though I would like the option to print units and maybe change the numeric
format (i.e. engineering vs. scientific vs. raw).