add splash screen class
I have added a crude implementation for a splash screen, in the file
splashscreen.py. Its implementation is based on two excellent recipes
from activestate.com, namely

http://code.activestate.com/recipes/534124-elegant-tkinter-splash-screen/
and
http://code.activestate.com/recipes/577271-tkinter-splash-screen/

My class is an amalgam of the two, in that it loads an image using
tkinter's PhotoImage class, but does not use the __enter__ and __exit__
magic methods.